Harden session handling by using random client session access tokens that rotate on refresh.
This refactor changes the use of the request_id as the client session access token (auth_token cookie value) and the keyval store key.  On keyval store set the new behavior generates a random value for the client session access token.  This value is then hashed using the oidc_hmac_key and the hashed value is used as the keyval store ID and as a session ID for the logs.  The hashed value is not redeemable for a session, hardening the config against attacks on the logs and keyval stores.  The client session access token is rotated each refresh, which was done both for security and because it allowed for code deduplication.