Change default keyval state file location. (#90)
Previously, the keyval state file was configured to be stored in the "conf.d"
directory. By default, the NGINX process does not have write access to this
directory, necessitating users to either specify a different path or alter
the directory permissions.

The default path for the state file has been changed to "/var/lib/nginx/state".
This new location is more suitable for most Linux users and aligns with
security best practices, as only the NGINX user has read and write permissions
by default.